requirements.txt
ruff==0.12.11

python-docx==1.2.0